% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phase_day.R
\name{calc_phase_day}
\alias{calc_phase_day}
\title{Determine the different phases of the day}
\usage{
calc_phase_day(
  dataset,
  light = "lightatsurf",
  date = "date",
  id = ".id",
  lon = "lon",
  lat = "lat",
  method = "noaa"
)
}
\arguments{
\item{dataset}{A dataset containing the light level per individual per time}

\item{light}{Name of the light-level column}

\item{date}{Name of the datetime column}

\item{id}{Name of the individual ID column}

\item{lon}{Name of the Longitude column}

\item{lat}{Name of the Latitude column}

\item{method}{Method used to identify phases of day (\code{"noaa"} or \code{"dbscan"})}
}
\value{
Return the same \code{data.table} with an additional \code{phase} columns
}
\description{
This function can estimate day-time and night-time based on
light level (\code{dbscan}) or using the location and time data (\code{noaa}).
}
\details{
Methods:
\describe{
\item{DBSCAN}{use dbscan clustering algorithm to identify night time and so daytime.}
\item{NOAA}{use the algorithm provided by the National Oceanic & Atmospheric Administration (NOAA).}
}
}
\examples{
\dontrun{
# load data
data_nes <- get_data("nes")

# night and day calculation
result <- calc_phase_day(rbindlist(data_nes$year_2018,
  use.name = TRUE,
  idcol = TRUE
))
}

}
\references{
\code{vignette("phase_of_the_day")}
}
\seealso{
\code{\link[maptools]{crepuscule}}
}
