% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dt.R
\name{check_dt}
\alias{check_dt}
\title{Check is an object is a `data.table`}
\usage{
check_dt(data)
}
\arguments{
\item{data}{}
}
\value{
A boolean:
TRUE: the object is a data.table
FALSE: the object is not a data.table
}
\description{
Allow to rapidly check if an object is a `data.table`
}
\examples{
# data.table creation
dt = data.table()

# check if `dt` is a data.table
check_dt(dt)
}
