% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_copernicus.R
\docType{data}
\name{data_cop}
\alias{data_cop}
\title{Oceanographic information from North East Pacific}
\format{
A data table with one row = one cell for a specific time and location
\describe{
\item{time}{date}
\item{latitude}{latitude}
\item{longitude}{longitude}
\item{uo}{eastward ocean current velocity (m/s)}
\item{vo}{northward ocean current velocity (m/s)}
\item{so}{salinity (psu)}
\item{thetao}{potential temperature (°C)}
\item{zos}{sea surface height above geoid (m)}
\item{vel}{ocean current velocity (m/s)}
}
}
\source{
\href{https://resources.marine.copernicus.eu/product-detail/GLOBAL_MULTIYEAR_PHY_001_030/INFORMATION}{https://resources.marine.copernicus.eu/product-detail/GLOBAL_MULTIYEAR_PHY_001_030/INFORMATION}
}
\usage{
data_cop
}
\description{
This data where retrieve from \href{copernicus}{https://resources.marine.copernicus.eu/}
using the product \strong{GLOBAL_MULTIYEAR_PHY_001_030}.
}
\keyword{datasets}
