% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_nes}
\alias{data_nes}
\title{Diving parameters of weanlings in 2016 and 2018}
\format{
A list containing two lists (one for each year) with the same amount
of datasets than the number of individuals
\describe{
\item{...}{...}
}
}
\source{
\email{roxanne@ucsc.edu}
}
\usage{
data_nes
}
\description{
A list of datasets containing diving parameters for:
\itemize{
\item 6 individuals in 2016
\item 4 individuals in 2018
}
}
\keyword{datasets}
