% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-n (conflicted copy 2022-01-12 154911).R,
%   R/dot-n.R
\name{.n}
\alias{.n}
\title{Retrieve elements' name of a list within apply function}
\usage{
.n()

.n()
}
\value{
The name of the element considered

The name of the element considered
}
\description{
This small function should be used within an `apply` function when retrieving the element's name is needed.

This small function should be used within an `apply` function when retrieving the element's name is needed.
}
\examples{
# list creation
A = list(name_1 = NULL, name_2 = NULL)

# print name of each element using .n()
sapply(A,function(x) .n())

# list creation
A = list(name_1 = NULL, name_2 = NULL)

# print name of each element using .n()
sapply(A,function(x) .n())

}
\references{
\href{https://stackoverflow.com/questions/9950144/access-lapply-index-names-inside-fun}{https://stackoverflow.com/questions/9950144/access-lapply-index-names-inside-fun}

\href{https://stackoverflow.com/questions/9950144/access-lapply-index-names-inside-fun}{https://stackoverflow.com/questions/9950144/access-lapply-index-names-inside-fun}
}
