% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_cor_mat.R
\name{flat_cor_mat}
\alias{flat_cor_mat}
\title{Flatten correlation matrix}
\usage{
flat_cor_mat(cormat, pmat)
}
\arguments{
\item{cormat}{A correlation matrix}

\item{pmat}{A p-value matrix}
}
\value{
A dataframe with correlation and p-values for each pairwise comparison
}
\description{
Flatten correlation matrix into a 2D dataframe based on \code{\link[http://www.sthda.com/french/wiki/matrice-de-correlation-formattage-et-visualisation]{http://www.sthda.com/french/wiki/matrice-de-correlation-formattage-et-visualisation}}
}
\examples{
# Dataset generation
data_test = data.frame(sapply(c(1:10),function(x) rnorm(10)))

# Correlation matrix calculation
cor_test = cor(data_test)

# Associated p-values
cor_test_p = ggcorrplot::cor_pmat(data_test)

# Flatten the correlation matrix
flat_cor_mat(cor_test, cor_test_p)

}
