% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Retrieve data}
\usage{
get_data(species = "all")
}
\arguments{
\item{species}{\code{"all"}, \code{"nes"} or \code{"ses"} to retrieve respectively the
dataset for both species or only the one associated with northern
or southern elephant seals.}
}
\value{
A table of class \code{data.table} if \code{species = "all"}, otherwise
a list by year and individual
}
\description{
This function will first look for datasets in the \code{ontodive} package. If datasets
are not included, it will then tried to download the dataset based on the DOI
pre-reserved by Dryad when submitting the paper. If somehow it still doesn't work
you will be asked to enter the DOI of the dataset you're looking for.
}
\examples{
\dontrun{
# retrieve all data
get_data("all")
}
}
