% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{Find outliers using interquartile range rule}
\usage{
is_outlier(x)
}
\arguments{
\item{x}{A vector in which outliers need to be found}
}
\value{
A boolean vector, TRUE for outliers, FALSE when not
}
\description{
This function aims at finding outliers in a vector using the interquartile range rule
}
\examples{
# generate data from gaussian and exponential distribution
X = sample(c(rnorm(100), c(rexp(50), -rexp(50))))
# plot data
plot(X, col = ifelse(is_outlier(X),"red","black"))#'
}
