% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_treatment.R
\name{location_treatment}
\alias{location_treatment}
\title{Treat location data with a continuous-time state-space model}
\usage{
location_treatment(
  data,
  model = "crw",
  time.step = 1,
  vmax = 3,
  with_plot = FALSE,
  export = NULL
)
}
\arguments{
\item{data}{Dataset of observation, usually the file \*Argos.csv or
\*Location.csv files}

\item{model}{Choose to fit either a simple random walk ("rw") or correlated
random walk ("crw") as a continuous-time process model}

\item{time.step}{options: 1) the regular time interval, in hours, to predict
to; 2) a vector of prediction times, possibly not regular, must be specified as a data.frame with id and POSIXt dates; 3) NA - turns off prediction and locations are only estimated at observation times.}

\item{vmax}{The max travel rate (m/s) passed to sda to identify outlier locations}

\item{with_plot}{A diagnostic plot}

\item{export}{To export the new generated dataset}
}
\value{
A dataset with the new location data
}
\description{
Using `fit_ssm` function from `foieGras` package, this function "clean" the
location data to be used for further analysis at the dive scale.
}
\examples{
# load library
library(foieGras)
library(data.table)

# run this function on sese1 dataset included in foieGras package
output <- location_treatment(copy(sese1), with_plot = TRUE)

}
\references{
run_foieGras_generic.R (\email{tkeates@ucsc.edu})

\href{https://ianjonsen.github.io/foieGras/}{https://ianjonsen.github.io/foieGras/}
}
\seealso{
\code{\link[foieGras]{fit_ssm}}
}
