% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_to_posix.R
\name{matlab_to_posix}
\alias{matlab_to_posix}
\title{matlab_to_posix}
\usage{
matlab_to_posix(x, timez = "UTC")
}
\arguments{
\item{x}{column with Matlab datenum}

\item{timez}{time zone to use}
}
\value{
A vector of POSIXct
}
\description{
matlab_to_posix
}
\examples{
matlab_to_posix(c(737182.4))
}
\references{
\code{\link["http://lukemiller.org/index.php/2011/02/converting-matlab-and-r-date-and-time-values/"]{http://lukemiller.org/index.php/2011/02/converting-matlab-and-r-date-and-time-values/}}
}
