% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp.R
\name{plot_comp}
\alias{plot_comp}
\title{Plot the evolution of a diving parameter accross time between two populations}
\usage{
plot_comp(
  data,
  diving_parameter = NULL,
  sp = "sp",
  time = "day_departure",
  id = ".id",
  nb_days = 100,
  bs = "cs",
  k = 6,
  alpha_point = 0.01
)
}
\arguments{
\item{data}{A dataset containing the required information}

\item{diving_parameter}{The colname associated with the diving parameter to represent}

\item{sp}{The colname associated with species}

\item{time}{The colname associated with time}

\item{id}{The colname associated with individual}

\item{nb_days}{How many days to represent}

\item{bs}{Smooth terms in GAM}

\item{k}{The dimension of the basis used to represent the smooth term}

\item{alpha_point}{The transparency of the point}
}
\value{
Return a ggplot
}
\description{
This function approximate the relationship between the considered diving
behavior and time in order to better represent this evolution with a smooth
curve, rather than scatterplot
}
\details{
This function fits a GAM with the species as a grouping factor and a random
effect (intercept + slope) on the individual (*i.e.* diving parameter ~
species + s(time) + (1 time | individual). This allows to represent a curve
for each species, but also to access to the curve associated with each
individual.
}
\examples{
\dontrun{
# load data
data("data_nes")
data("data_ses")

# combine
data_comp <- rbind(
  rbindlist(data_nes$year_2018) \%>\%
    .[, sp := "nes"],
  rbindlist(data_ses$year_2014) \%>\%
    .[, sp := "ses"],
  use.names = T,
  fill = T
)

# plot
plot_comp(data_comp, "maxdepth")
}

}
\references{
\href{https://stats.stackexchange.com/questions/403772/different-ways-of-modelling-interactions-between-continuous-and-categorical-pred}{https://stats.stackexchange.com/questions/403772/different-ways-of-modelling-interactions-between-continuous-and-categorical-pred}

James Grecian. (2022). jamesgrecian/harpPup: v1.0 (v1.0). Zenodo. \href{https://doi.org/10.5281/zenodo.5901391}{https://doi.org/10.5281/zenodo.5901391}
}
\seealso{
\code{\link[mgcv]{smooth.terms}}

\code{\link[mgcv]{gam}}
}
