% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jjo.R
\name{theme_jjo}
\alias{theme_jjo}
\title{A personnalised ggplot theme}
\usage{
theme_jjo(base_size = 12)
}
\arguments{
\item{base_size}{Font size}
}
\value{
a ggplot theme
}
\description{
A personnalised ggplot theme
}
\examples{
# load package
library(ggplot2)

# plot
ggplot(cars) +
  geom_point(aes(x = speed, y = dist)) +
  theme_jjo()
}
\references{
\href{https://benjaminlouis-stat.fr/en/blog/2020-05-21-astuces-ggplot-rmarkdown/}{https://benjaminlouis-stat.fr/en/blog/2020-05-21-astuces-ggplot-rmarkdown/}
}
