% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_github_vignette.R
\name{build_github_vignette}
\alias{build_github_vignette}
\title{Build Vignette for Github}
\usage{
build_github_vignette()
}
\value{

}
\description{
This function aims at processing vignettes (building, organizing) before pushing on Github so that user can download a package with vignettes
}
\examples{
build_github_vignette()
}
\references{
\code{\link[community.rstudio.com/t/how-to-add-vignette-html-or-r-files-to-a-github-rep/45905/7]{https://community.rstudio.com/t/how-to-add-vignette-html-or-r-files-to-a-github-rep/45905/7}}
}
