% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phase_day.R
\name{calc_phase_day}
\alias{calc_phase_day}
\title{Determine the different phasesof the day}
\usage{
calc_phase_day(dataset, method = "dbscan")
}
\arguments{
\item{dataset}{A dataset containing the light level per individual per time}

\item{method}{Method used to identify phases of day}
}
\value{
Return the same data.table with an additional "phase" columns
}
\description{
This function can estimate daytime and nightime based on light level using several different methods.
}
\details{
Methods

* DBSCAN: use the dbscan clustering algorithm to identify night time and so daytime.

* ...: to be implemented
}
\examples{
\dontrun{
# load data
data("data_nes")

# night and day calculation
result <- calc_phase_day(rbindlist(
}

}
