% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dt.R
\name{check_dt}
\alias{check_dt}
\title{Check if an object is a `data.table`}
\usage{
check_dt(dt)
}
\arguments{
\item{dt}{The object to check}
}
\value{
A boolean:

* TRUE the object is a data.table

* FALSE the object is not a data.table
}
\description{
Allow to quickly check if an object is a `data.table`
}
\details{
This function checks if the class `data.table` is among the considered object's classes.
}
\examples{
# load data.table
library(data.table)

# data.table creation
dt <- data.table()

# check if `dt` is a data.table
check_dt(dt)
}
