% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_copernicus.R
\docType{data}
\name{data_cop}
\alias{data_cop}
\title{Oceanographic information from North East Pacific and Southern Ocean}
\format{
A list of two data tables, one for northern elephant seals and
the other for southern elephant seals, where each row correspond to a
cell of 0.083° × 0.083° for a specific time and location.
\itemize{
\item $northen: [lat=c(25, 55); lon=c(-165, -115); time c("2018-05-01", "2019-05-01")]
\item $southern: [lat=c(-65, -40); lon=c(35, 115); time c("2015-01-01", "2015-11-30")]
}

\describe{
\item{time}{date}
\item{latitude}{latitude}
\item{longitude}{longitude}
\item{uo}{eastward ocean current velocity (m/s)}
\item{vo}{northward ocean current velocity (m/s)}
\item{so}{salinity (psu)}
\item{thetao}{potential temperature (°C)}
\item{zos}{sea surface height above geoid (m)}
\item{vel}{ocean current velocity (m/s)}
}
}
\source{
\href{https://resources.marine.copernicus.eu/product-detail/GLOBAL_MULTIYEAR_PHY_001_030/INFORMATION}{https://resources.marine.copernicus.eu/product-detail/GLOBAL_MULTIYEAR_PHY_001_030/INFORMATION}
}
\usage{
data(data_cop)
}
\description{
This dataset was retrieved from
\href{https://resources.marine.copernicus.eu/}{Copernicus} using the product
\strong{GLOBAL_MULTIYEAR_PHY_001_030} and the dataset
\strong{cmems_mod_glo_phy_my_0.083_P1D-m}.
}
\keyword{datasets}
