% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-n.R
\name{.n}
\alias{.n}
\title{Retrieve elements' name of a list within apply function}
\usage{
.n()
}
\value{
The name of the element considered
}
\description{
This small function should be used within an apply function when retrieving the element's name is needed.
}
\examples{
A = list(name_1 = NULL, name_2 = NULL)
sapply(A,function(x) .n())
}
\references{
<https://stackoverflow.com/questions/9950144/access-lapply-index-names-inside-fun>
}
