% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_col.R
\name{format_col}
\alias{format_col}
\title{Format}
\usage{
format_col(col_names)
}
\arguments{
\item{col_names}{A vector of character containing column names to be
formatted}
}
\value{
The input correctly formatted
}
\description{
Format columns of a dataset by removing capital letters, bracket and
transform white space as underscore
}
\examples{
# format several vectors of characters
format_col(c("lkjez alzkj","kjfe[lkjz]","lkjee [lkj]"))

}
