% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ind.R
\name{plot_ind}
\alias{plot_ind}
\title{Display summary information for a seal}
\usage{
plot_ind(data_seal, ind, col_text = "black", col_back = "transparent")
}
\arguments{
\item{data_seal}{A dataset containing information on the dives (maxdepth, lat, lon, date, driftrate)}

\item{ind}{The name of the individual considered}

\item{col_text}{The color of the text}

\item{col_back}{The color of the background}
}
\value{
A ggplot object which contains all figures
}
\description{
This function displays several key information regarding a seal.
}
\details{
* A map of the seal's trip at sea with a color code based on the number of days since departure

* The max depth reached for each dives across time

* The evolution of their daily median drift rate across time

* the evolution of their ADL across time
}
\examples{
# load data
 data("data_nes")

 # plot result
 plot_ind(data_nes$year_2018$ind_2018070, ind = "ind_2018070")

}
